% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diatoms_galapagos}
\alias{diatoms_galapagos}
\title{Diatoms - Seddon et al. (YYYY)}
\format{
An object of class S3 (tibble) with dimensions 6,696 × 4.
}
\source{
\href{url or doi here}{citation or poublisher whatever}
}
\usage{
data(seddondiatoms)
}
\description{
Data from publication....more description here.
}
\examples{
data(diatoms_galapagos)
}
\references{
(\href{url or doi here}{publisher})
}
\keyword{dataset,}
\keyword{diatom}
\keyword{regime}
\keyword{shift,}
