% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{everglades}
\alias{everglades}
\title{Everglades data title will go here}
\format{
An object of class S3 (tibble) with dimensions 20,664 × 4.
}
\source{
\href{url or doi here}{citation or publisher here}
}
\usage{
data(everglades)
}
\description{
Data from publication....more description here.
}
\examples{
data(everglades)
}
\references{
reference goes here
(\href{url or doi here}{publisher})
}
\keyword{abrupt}
\keyword{change}
\keyword{community,}
\keyword{dataset,}
\keyword{everglades,}
