% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire_konza_ratajczak}
\alias{fire_konza_ratajczak}
\title{Konza LTER Fires - Ratajczak et al.}
\format{
A list of two elements: c('dataset','attributes'). Dataset contains the original data. Attributes describe the use of each variable in the paper referenced.
}
\source{
\href{url or doi here}{citation or poublisher whatever}
}
\usage{
data(konza_fire)
}
\description{
Konza LTER Fire Dataset
}
\examples{
data(fire_konza_ratajczak)
}
\references{
(\href{url or doi here}{publisher})
}
\keyword{burn}
\keyword{controlled}
\keyword{dataset,}
\keyword{fire,}
\keyword{lter,}
\keyword{regime}
\keyword{shift,}
