% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{code}
\name{gradual_noncrit}
\alias{gradual_noncrit}
\title{Simulated Data Exhibiting Gradual Non-Critical Shift}
\format{
.R  document
}
\source{
\href{url or doi here}{citation or publisher here}
}
\usage{
code(stegnersimulation)
}
\description{
Code adapted from Stegner et al (2019) Inferring critical transitions in paleoecological time series with irregular sampling and variable time-averaging. Quaternary Science Reviews 207: 49-63 DOI: 10.1016/j.quascirev.2019.01.009
}
\examples{
data(gradual_noncrit)
}
\references{
Stegner, Ratajczak, Carpenter, & Williams (2019) Inferring critical transitions in paleoecological time series with irregular sampling and variable time-averaging. Quaternary Science Reviews 207: 49-63
(\href{10.1016/j.quascirev.2019.01.009}{publisher})
}
\keyword{code,}
\keyword{cover}
\keyword{simulation,}
\keyword{woody}
