% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundfish}
\alias{groundfish}
\title{Groundfish data title}
\format{
An object of class XXX.
}
\source{
\href{url or doi here}{citation or poublisher whatever}
}
\usage{
data(groundfish)
}
\description{
Description here
here
here
}
\examples{
data(groundfish)
}
\references{
Pedersen et al. something or other
(\href{url or doi here}{publisher})
}
\keyword{datasets}
