% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_dataset.R
\name{munge_dataset}
\alias{munge_dataset}
\title{Munge Datasets to Conform to Package Standards}
\usage{
munge_dataset(data, attributes, df.citation = NULL)
}
\arguments{
\item{data}{The dataset being created}

\item{attributes}{The attributes table describing the variables existing in *data*}

\item{df.citation}{Default NULL. Ideally a string containing the citation for the dataset in use in the style of .bibtex.}
}
\description{
A function for munging each dataset accordin gto what is specified in teh attribues table
}
