% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_change}
\alias{no_change}
\title{Simulated Data for Data Exhibiting No Transition}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 3 columns.
}
\source{
\href{url or doi here}{citation or publisher here}
}
\usage{
no_change
}
\description{
Code adapted from Stegner et al (2019) Inferring critical transitions in paleoecological time series with irregular sampling and variable time-averaging. Quaternary Science Reviews 207: 49-63 DOI: 10.1016/j.quascirev.2019.01.009
}
\examples{
data(no_change)
}
\references{
Stegner, Ratajczak, Carpenter, & Williams (2019) Inferring critical transitions in paleoecological time series with irregular sampling and variable time-averaging. Quaternary Science Reviews 207: 49-63
(\href{10.1016/j.quascirev.2019.01.009}{publisher})
}
\keyword{cover}
\keyword{simulation,}
\keyword{woody}
