% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seddondiatoms}
\alias{seddondiatoms}
\title{Seddon Diatoms data}
\format{
An object of class S3 (tibble) with dimensions 6,696 × 4.
}
\source{
\href{url or doi here}{citation or poublisher whatever}
}
\usage{
data(seddondiatoms)
}
\description{
Data from publication....more description here.
}
\examples{
data(seddondiatoms)
}
\references{
(\href{url or doi here}{publisher})
}
\keyword{datasets,}
\keyword{diatoms}
\keyword{regime}
\keyword{shift,}
