% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundfish}
\alias{groundfish}
\title{Groundfish data title will go here}
\format{
An object of class S3 (tibble) with dimensions 105,879 × 9.
}
\source{
\href{url or doi here}{citation or publisher here}
}
\usage{
data(groundfish)
}
\description{
Data from publication....more description here.
}
\examples{
data(groundfish)
}
\references{
Pedersen et al. something or other
(\href{url or doi here}{publisher})
}
\keyword{abrupt}
\keyword{change}
\keyword{community,}
\keyword{dataset,}
\keyword{fish,}
\keyword{regime}
\keyword{shift,}
